Refactor: Refactor directory struct
- src/Frouros/Net/Models/LxProto.cs: Seperate structs in each file
- src/Frouros/Net/Models/NativePacketLog.Consts.cs: Seperate structs in each file
- src/Frouros/Net/Models/NativePacketLog.cs: Seperate structs in each file
- src/Frouros/Net/Models/PacketLog.cs: Seperate structs in each file
- src/Frouros/Net/PacketLog.cs: Merge into Models folder
- src/Frouros/Utils/Endianness.cs: Merge into Utils folder
